+++ custom_components/vn_calendar_component/const.py
@@ -4,7 +4,7 @@ DEFAULT_NAME = "VN Lunar Calendar Component"
 
 TIME_ZONE = 7
 
-VERSION = "1.0.0"
+VERSION = "1.1.0"
 
 # ==== SERVICES ====
 SERVICE_GET_DAY = "get_day"
@@ -12,6 +12,8 @@ SERVICE_GET_MONTH = "get_month"
 SERVICE_GET_YEAR = "get_year"
 SERVICE_TODAY = "today"
 SERVICE_CLEANUP = "cleanup"
+SERVICE_GOOD_HOURS = "good_hours"
+SERVICE_GOOD_HOURS_TODAY = "good_hours_today"
 
 # ==== PARAMS ====
 PARAM_DAY = "day"
@@ -19,8 +21,17 @@ PARAM_MONTH = "month"
 PARAM_YEAR = "year"
 
 # ==== ENTITY ====
-BSENSOR_VEGDAY_UNIQUE_NAME = "VN Calendar Veg Day"
+BSENSOR_VEGDAY_UNIQUE_NAME = "VN Calendar veg day"
 BSENSOR_VEGDAY_UNIQUE_ID = "vn_calendar_veg_day"
 
-SENSOR_LUNARTODAY_UNIQUE_NAME = "VN Calendar Today"
+BSENSOR_GOODHOUR_UNIQUE_NAME = "VN Calendar good hour"
+BSENSOR_GOODHOUR_UNIQUE_ID = "vn_calendar_good_hour"
+
+SENSOR_LUNARTODAY_UNIQUE_NAME = "VN Calendar today"
 SENSOR_LUNARTODAY_UNIQUE_ID = "vn_calendar_today"
+
+SENSOR_DAYTYPE_UNIQUE_NAME = "VN Calendar day type"
+SENSOR_DAYTYPE_UNIQUE_ID = "vn_calendar_day_type"
+
+SENSOR_SOLARTERM_UNIQUE_NAME = "VN Calendar solar term"
+SENSOR_SOLARTERM_UNIQUE_ID = "vn_calendar_solar_term"